Moved initialization check from beamcut_reduction_pipeline.py to utils/user_interaction.py.
